requirements.txt
fastapi
openai
python-dotenv
python-multipart
sqlmodel
